% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectedPathCost.r
\name{ExpectedPathCost}
\alias{ExpectedPathCost}
\title{Expected path costs}
\usage{
ExpectedPathCost(
  x,
  ODdemand,
  ODpair,
  A = diag(length(x)),
  Alpha,
  Beta,
  pow = 4,
  prob.model = "ProductMN"
)
}
\arguments{
\item{x}{Vector of mean link flows}

\item{ODdemand}{Vector of origin-destination (OD) travel demands}

\item{ODpair}{Vector indicating the OD pair serviced by each route (ordered by columns of the path-link incidence matrix, A).}

\item{A}{Path-link incidence matrix}

\item{Alpha}{Vector of free flow travel time parameters for each link}

\item{Beta}{Vector of capacity parameters for each link}

\item{pow}{Polynomial order of cost function for each link. Defaults to 4.}

\item{prob.model}{Route choice probability model. One of "ProductMN" (product multinomial, the default), or "Poisson".}

\item{theta}{A dispersion parameter. For the logit model, a single value specifying the logit parameter. For the probit model, a vector of standard deviations for the individual link cost errors. Defaults to 1.}
}
\value{
Vector of expected path costs
}
\description{
This function calculates expected path costs as a function of mean link flows under various probability models. Its primary use is when called by the function GSUE. For the product multinomial model, a normal approximation to higher order moments is applied unless the link path incidencde matrix is diagonal.
}
\examples{
ExpectedPathCost()
}
\keyword{costs}
\keyword{expected}
\keyword{mean}
\keyword{path}
