% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSUE.r
\name{GSUE}
\alias{GSUE}
\title{Generalized Stochastic User Equilibrium}
\usage{
GSUE(
  ODdemand,
  ODpair,
  A,
  Alpha,
  Beta,
  pow = 4,
  RUM = "logit",
  theta = 1,
  prob.model = "ProductMN",
  tol = 1e-04,
  verbose = F
)
}
\arguments{
\item{ODdemand}{Vector of origin-destination (OD) travel demands}

\item{ODpair}{Vector indicating the OD pair serviced by each route (ordered by columns of the path-link incidence matrix, A).}

\item{A}{Path-link incidence matrix}

\item{Alpha}{Vector of free flow travel time parameters for each link}

\item{Beta}{Vector of capacity parameters for each link}

\item{pow}{Polynomial order of cost function for each link. Defaults to 4.}

\item{RUM}{Choice of random utility model. Can be "logit" (the default) or "probit".}

\item{theta}{A dispersion parameter. For the logit model, a single value specifying the logit parameter. For the probit model, a vector of standard deviations for the individual link cost errors. Defaults to 1.}

\item{prob.model}{Route choice probability model. One of "ProductMN" (product multinomial, the default), or "Poisson".}

\item{tol}{Tolerance for convergence assessment (measured as route mean squared difference between current flow vector and the search direction). Defaults of 1e-4.}

\item{verbose}{Should progress of algorithm be printed out? Defaults to FALSE.}
}
\value{
Output is a list of SUE route flows x and corresponding path costs u.
}
\description{
This function calculates generalized stochastic user equilibrium (GSUE) using the method of successive averages.
}
\examples{
A <- matrix(c(0,1,0,0,1,0,0,0,0,1,1,0,0,0,1,0,0,0,0,1,1,0,0,0,0,0,0,1),ncol=4,byrow=T) 
Alpha <- rep(10,7)
Beta <- rep(50,7)
pow <- rep(4,7)
ODpair <- c(1,1,2,2)
ODdemand <- c(50,50)
theta <- 0.7
GSUE(ODdemand,ODpair,A=A,Alpha=Alpha,Beta=Beta,pow=pow,theta=theta,verbose=F)
}
\keyword{GSUE}
