% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-diag.r
\name{is_diag}
\alias{is_diag}
\title{Check if matrix is diagonal}
\usage{
is_diag(A)
}
\arguments{
\item{A}{Matrix to be tested}
}
\value{
Logical value indicating whether matrix is diagonal
}
\description{
This function indicates whether the matrix supplied is diagonal.
}
\examples{
A <- cbind(c(1,1),c(0,1)
is_matrix(A)
}
\keyword{diagonal}
\keyword{matrix}
