% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitPathProb.r
\name{LogitPathProb}
\alias{LogitPathProb}
\title{Probability of path choices using logit model.}
\usage{
LogitPathProb(u, ODpair, theta = 1)
}
\arguments{
\item{u}{Vector of utlities on each path}

\item{ODpair}{Vector indicating the OD pair serviced by each route.}

\item{theta}{The logit parameter. Defaults to 1.}
}
\value{
Vector of route choice probabilities
}
\description{
This function calculates probabilities for selection of each path on the network based on set of path disutilities (costs). The function uses a logit random utility model.
}
\examples{
A <- matrix(c(0,1,0,0,1,0,0,0,0,1,1,0,0,0,1,0,0,0,0,1,1,0,0,0,0,0,0,1),ncol=4,byrow=T)
ODpair <- c(1,1,2,2)
Alpha <- rep(10,7)
Beta <- rep(2,7)
pow <- rep(4,7)
path_flow <- c(10,20,15,15)
x <- A\%*\%path_flow
u <- PathCost(x,A,Alpha,Beta)
LogitPathProb(u,ODpair,theta=0.7)
}
\keyword{path}
\keyword{probability}
\keyword{route}
