% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathProb.r
\name{PathProb}
\alias{PathProb}
\title{Probability of path choices}
\usage{
PathProb(u, ODpair, A, RUM = "logit", theta = 1)
}
\arguments{
\item{u}{Vector of utlities on each path}

\item{ODpair}{Vector indicating the OD pair serviced by each route (ordered by columns of the path-link incidence matrix, A).}

\item{A}{Path-link incidence matrix}

\item{RUM}{Choice of random utility model. Can be "logit" (the default) or "probit".}

\item{theta}{A dispersion parameter. For the logit model, a single value specifying the logit parameter. For the probit model, a vector of standard deviations for the individual link cost errors. Defaults to 1.}
}
\value{
Vector of route choice probabilities
}
\description{
This function calculates probabilities for selection of each path on the network based on set of path disutilities (costs). The function uses either a logit or probit random utility model.
}
\examples{
A <- matrix(c(0,1,0,0,1,0,0,0,0,1,1,0,0,0,1,0,0,0,0,1,1,0,0,0,0,0,0,1),ncol=4,byrow=T)
ODpair <- c(1,1,2,2)
Alpha <- rep(10,7)
Beta <- rep(2,7)
pow <- rep(4,7)
path_flow <- c(10,20,15,15)
x <- A\%*\%path_flow
u <- PathCost(x,A,Alpha,Beta)
PathProb(u,ODpair,A,RUM="logit",theta=0.7)
}
\keyword{path}
\keyword{probability}
\keyword{route}
