% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbitPathProb.r
\name{ProbitPathProb}
\alias{ProbitPathProb}
\title{Probability of path choices using probit model.}
\usage{
ProbitPathProb(u, ODpair, A, theta, Niter = 1000)
}
\arguments{
\item{u}{Vector of utlities on each path}

\item{ODpair}{Vector indicating the OD pair serviced by each route.}

\item{A}{Path-link incidence matrix}

\item{theta}{Vector of standard deviations for link cost errors. Defaults to 1 in each case.}

\item{Niter}{Number of iterations when computing probabilities by simulation. Defaults to 1000.}
}
\value{
Vector of route choice probabilities
}
\description{
This function calculates probabilities for selection of each path on the network based on set of path disutilities (costs). The function uses a probit random utility model. Calculations are exact for two route networks. In other case the probabilities are approximated by simulation.
}
\examples{
A <- diag(2)
ODpair <- c(1,1)
u <- c(15,17)
LogitPathProb(u,ODpair,A,theta=1)
}
\keyword{path}
\keyword{probability}
\keyword{route}
